+++ core/src/iAModality.h
@@ -61,9 +61,9 @@ public:
 	//! return the channel in the specified file that the data in this class comes from (don't confuse with channelID!)
 	int channel() const;
 	//! return the number of components in this modality
-	int componentCount() const;
+	size_t componentCount() const;
 	//! return a specific component of this modality
-	vtkSmartPointer<vtkImageData> component(int idx) const;
+	vtkSmartPointer<vtkImageData> component(size_t idx) const;
 	//! get the name of the transfer function file
 	QString transferFileName() const;
 	//! set name of the modality
